+++ packages/create-block/lib/scaffold.js
@@ -26,6 +26,7 @@ module.exports = async (
 		description,
 		dashicon,
 		category,
+		textdomain,
 		attributes,
 		supports,
 		author,
@@ -95,7 +96,7 @@ module.exports = async (
 		customPackageJSON,
 		customBlockJSON,
 		example,
-		textdomain: slug,
+		textdomain: textdomain || slug,
 		rootDirectory,
 	} );